feat: move landing page to another folder
