+++ src/App.tsx
@@ -1,6 +1,6 @@
 
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import LandingPage from './components/pages/LandingPage';
+import LandingPage from './pages/LandingPage';
 
 function App() {
   return (
